+++ R/zzz.R
@@ -1,3 +1,5 @@
+.packageExclude <- c("MPI", "PSOCK")
+
 ## covr: skip=all
 .onLoad <- function(libname, pkgname) {
   debug <- isTRUE(as.logical(getEnvVar2("R_PARALLELLY_DEBUG", "FALSE")))
@@ -56,6 +58,8 @@
     envs <- sprintf("- %s=%s", names(envs), sQuote(envs))
     mdebug(paste(c("parallelly-specific environment variables:", envs), collapse = "\n"))
   }
+
+  registerClusterTypes()
 }