+++ src/000.api.h
@@ -1,2 +1,3 @@
 /* C-level API that is called from R */
 SEXP R_test_tcp_port(SEXP port);
+SEXP R_calc_serialized_size(SEXP robj);