+++ src/calc-serialized-size.h
@@ -1,0 +1,2 @@
+R_xlen_t R_calc_serialized_size(SEXP robj);
+